Separate chart module. Low prices are displayed better
