+++ lib/sanbase/notifications/price_volume_diff.ex
@@ -167,7 +167,7 @@ defmodule Sanbase.Notifications.PriceVolumeDiff do
   end
 
   defp notification_embeds(project) do
-    Discord.build_embedded_chart(
+    Sanbase.Chart.build_embedded_chart(
       project,
       Timex.shift(Timex.now(), days: -90),
       Timex.shift(Timex.now(), days: -1)
